backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
web3==6.11.3
qrcode[pil]==7.4.2
Pillow==10.1.0
python-multipart==0.0.6
aiofiles==23.2.1
